Open Space conversations on separate route
