+++ web/apps/space/public/_redirects
@@ -1,2 +1,3 @@
 /app/friends/*    /app/friends/[spaceId]    200
+/app/messages/*    /app/messages/[spaceId]    200
 /*    /index.html    200
